Avoid some timeouts to work around keyboard handling

+++ app/js/file.js
@@ -58,48 +58,42 @@ define(function(require, exports, module) {
 
     command.define("File:Delete", {
         exec: function(edit) {
-            setTimeout(function() {
-                confirmDelete(edit);
-            }, 0);
+            confirmDelete(edit);
         }
     });
 
     command.define("File:Rename", {
         exec: function(edit) {
-            setTimeout(function() {
-                renameFile(edit);
-            }, 0);
+            renameFile(edit);
         }
     });
 
     command.define("File:Delete Tree", {
         exec: function() {
-            setTimeout(function() {
-                ui.prompt({
-                    message: "Prefix of tree to delete:",
-                    input: ""
-                }, function(err, prefix) {
-                    if (prefix) {
-                        ui.prompt({
-                            message: "Are you sure you want to delete all files under " + prefix + "?"
-                        }, function(err, yes) {
-                            if(!yes) {
-                                return;
-                            }
-                            project.listFiles(function(err, files) {
-                                files = _.filter(files, function(path) {
-                                    return path.indexOf(prefix) === 0;
-                                });
-                                async.each(files, function(path, next) {
-                                    project.deleteFile(path, next);
-                                }, function() {
-                                    goto.fetchFileList();
-                                    console.log("All files under", prefix, "removed!");
-                                });
+            ui.prompt({
+                message: "Prefix of tree to delete:",
+                input: ""
+            }, function(err, prefix) {
+                if (prefix) {
+                    ui.prompt({
+                        message: "Are you sure you want to delete all files under " + prefix + "?"
+                    }, function(err, yes) {
+                        if (!yes) {
+                            return;
+                        }
+                        project.listFiles(function(err, files) {
+                            files = _.filter(files, function(path) {
+                                return path.indexOf(prefix) === 0;
+                            });
+                            async.each(files, function(path, next) {
+                                project.deleteFile(path, next);
+                            }, function() {
+                                goto.fetchFileList();
+                                console.log("All files under", prefix, "removed!");
                             });
                         });
-                    }
-                });
+                    });
+                }
             });
         },
         readOnly: true
